% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pancan12_datasets.R
\docType{data}
\name{pancan12}
\alias{clinical.cb}
\alias{cnv.cb}
\alias{cnv.cb,}
\alias{expression.cb,}
\alias{expression.cb1}
\alias{expression.cb2}
\alias{mutation.cb}
\alias{mutation.cb,}
\alias{pancan12}
\title{PANCAN12 datasets from Cancer Browser project}
\source{
\url{https://genome-cancer.ucsc.edu/proj/site/hgHeatmap/}
}
\usage{
clinical.cb

cnv.cb

mutation.cb

expression.cb1

expression.cb2
}
\value{
Data frames PANCAN12 data.
}
\description{
Package provides clinical, expression, cnv and mutation data
from Genome Cancer Browser for 12 cohorts types. Converting
those datasets from \code{data.frames} to Bioconductor
classes is explained here \link{convertTCGA}.
}
\keyword{datasets}

