% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pancan12_datasets.R
\docType{data}
\name{pancan12}
\alias{clinical.cb}
\alias{cnv.cb}
\alias{cnv.cb,}
\alias{expression.cb,}
\alias{expression.cb1}
\alias{expression.cb2}
\alias{mutation.cb}
\alias{mutation.cb,}
\alias{pancan12}
\title{PANCAN12 datasets from Cancer Browser project}
\source{
\url{https://genome-cancer.ucsc.edu/proj/site/hgHeatmap/}
}
\usage{
clinical.cb

cnv.cb

mutation.cb

expression.cb1

expression.cb2
}
\description{
Package provides cnv datasets from Cancer Browser for 12
cohorts types
}
\keyword{datasets}

